// GC-Mobile-Web-Application/UnitedAdminPortal/src/app/models/event.model.ts
export class Event {
    EventID: string;
    Title: string;
    Description: string;
    StartDate: string;
    EndDate: string;
    Location: string;
    Prize: string;
    Type: string;
    ImageUrl: string;
    Tags: string;
    notRepeat: string;
    startTime: string;
    endTime: string;
    relatedStyle1: string;
    relatedStyle2: string;
    relatedStyle3: string;
    viewCounter: number;
    country: string;
    EventProUser: string;
    SeriesEndDate: string;
}